pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "holdup"
version = "0.0.1"
authors = [
  { name="David Cyze -- other authors please add yourself", email="dcyze3@gatech.edu" },
]
description = "Our final project"
readme = "README.md"
license = { file="LICENSE" }
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
    "numpy==1.23.1",
    "pandas==1.4.3",
    "scikit-learn==1.1.2",
    "pyyaml==6.0"
]

[project.urls]
"Homepage" = "https://github.com/pypa/sampleproject"
"Bug Tracker" = "https://github.com/pypa/sampleproject/issues"

[metadata]
name = "holdup"

[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
]
pythonpath = [
  ".", "src"
]